utilize f16 for dct+quant
DCT+Q about 200us faster

+++ c63.h
@@ -1,6 +1,7 @@
 #ifndef C63_C63_H_
 #define C63_C63_H_
 
+#include <arm_neon.h>
 #include <inttypes.h>
 #include <stdint.h>
 #include <stdio.h>
@@ -100,7 +101,7 @@ struct c63_common
   int me_search_range;
 
   uint8_t quanttbl[COLOR_COMPONENTS][64];
-  float quant_scale[COLOR_COMPONENTS][64]; // Scaled dct 0.25*quant_values, drop div in hot-loop
+  float16_t quant_scale[COLOR_COMPONENTS][64]; // Scaled dct 0.25*quant_values, drop div in hot-loop
 
   struct frame *refframe;
   struct frame *curframe;
